norm conform shunting yard utils and added shunting print file
