+++ Makefile
@@ -6,6 +6,8 @@ HEADERS	:= -I ./include
 
 SRCDIR := ./src/
 
+# SRCS := $(shell find $(SRCDIR) -name '*.c')
+
 SRCS	:=  $(SRCDIR)main.c \
 			$(SRCDIR)input.c \
 			$(SRCDIR)utils/split.c \
@@ -44,6 +46,7 @@ SRCS	:=  $(SRCDIR)main.c \
 			$(SRCDIR)shunting_yard/shunting_yard.c \
 			$(SRCDIR)shunting_yard/shunting_yard_utils.c \
 			$(SRCDIR)shunting_yard/shunting_stack_utils.c \
+			$(SRCDIR)shunting_yard/shunting_print.c \
 			$(SRCDIR)shunting_yard/ast_utils.c
 
 OBJDIR := ./obj/
